+++ shared/consts.ts
@@ -1,2 +1,4 @@
 export const networkPort = 7523;
 export const discordAppId = "1499508916722077777";
+export const steamProfilesUrl = "https://steamcommunity.com/profiles/";
+export const steamVanityUrl = "https://steamcommunity.com/id/";