+++ shared/util.ts
@@ -1,0 +1,4 @@
+export function isStringNumber(string: string) {
+    const num = Number(string);
+    return !isNaN(num) && isFinite(num);
+}