+++ .github/release-notes.md
@@ -1,2 +1,3 @@
 * Fixed error if Discord is not open when using custom RPC
 * Fixed list of past profile pictures overflowing
+* Added ability to search player profile url